Several admin verbs now use browser list input (#2420)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

This PR does the following:

**Browser list inputs**
- Slightly increases default height of the window
- Improves search text clarity

**Jump to Area, Jump To Mob, Get Mob, Send Mob, Jump to Key and Get Key
verbs**

- Now use browser list input instead of a default input, allowing them
to be searched and filtered through

**Example image:**

<img width="349" height="378" alt="NewSearch"
src="https://github.com/user-attachments/assets/3d959aa1-01b2-4b01-91fc-d27321daff7d"
/>

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Better input selection method for some verbs and improvements to browser
inputs themselves.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->